/* !!!Important - pay attention to terms like "serialized" used in Java Lang. I will subsequently create branches for the PYTHON version of this App. */
====================================
This commit is made for the driver's class.
I created a new branch so as to retain previous versions.

The next commit will be done with a new named branch. The branch and commit will have app; enable users save, and edit their diary's title, body and delete files. I will create a database class (preferably Sqlite) so that these functionalities can be effective.

I plan to get to this level also in python before the end of this week.

God bless Ya! peace